requirements.txt
# Core Data Science
pandas
numpy
openpyxl

# Firebase / GCP
firebase-admin
google-cloud-firestore
google-cloud-storage

# Visualization
matplotlib
seaborn
plotly

# Machine Learning & Modeling
scikit-learn
scipy

# Dashboard & Deployment
streamlit>=1.40.0
streamlit-option-menu
protobuf<5.0.0

# Development & EDA
ipykernel
jupyter
python-dotenv
requests
python-docx
google-generativeai
groq
Pillow
